[android] Moved SectionPosition to a separate class.
