+++ android/src/com/mapswithme/maps/bookmarks/BookmarkListAdapter.java
@@ -47,40 +47,6 @@ public class BookmarkListAdapter extends RecyclerView.Adapter<Holders.BaseBookma
   @Nullable
   private RecyclerLongClickListener mLongClickListener;
 
-  static class SectionPosition
-  {
-    static final int INVALID_POSITION = -1;
-
-    private final int mSectionIndex;
-    private final int mItemIndex;
-
-    SectionPosition(int sectionInd, int itemInd)
-    {
-      mSectionIndex = sectionInd;
-      mItemIndex = itemInd;
-    }
-
-    int getSectionIndex()
-    {
-      return mSectionIndex;
-    }
-
-    int getItemIndex()
-    {
-      return mItemIndex;
-    }
-
-    boolean isTitlePosition()
-    {
-      return mSectionIndex != INVALID_POSITION && mItemIndex == INVALID_POSITION;
-    }
-
-    boolean isItemPosition()
-    {
-      return mSectionIndex != INVALID_POSITION && mItemIndex != INVALID_POSITION;
-    }
-  }
-
   public static abstract class SectionsDataSource
   {
     @NonNull
